updating accuaracy

+++ Project1Lib/Game.h
@@ -162,7 +162,7 @@ public:
 
     std::shared_ptr<Sound> FindSoundByName(const wxString &audioName);
 
-    int CalculateAccuracy();
+    double CalculateAccuracy();
     void StopSound(char key);
 };
